+++ src/AssistanceTouristique/ServiceStatSiteOperations.java
@@ -10,6 +10,6 @@ public interface ServiceStatSiteOperations
     /**
      * Operation getStatsSite
      */
-    public AssistanceTouristique.Statistique[] getStatsSite(String date, int idSite);
+    public AssistanceTouristique.Statistique[] getStatsSite(String date, short idSite);
 
 }
