+++ src/Office/ServeurOffice.java
@@ -32,10 +32,10 @@ public class ServeurOffice {
     public static void main(String args[]) {
         //Tableau des id/nom des sites
         //Tableau des id/nom des sites 
-        HashMap<Integer,String> listeSites = new HashMap<Integer, String>();
-        listeSites. put(1, "Georges Labit");
-        listeSites.put(2,"Museum histoire naturelle");
-        listeSites.put(3, "Saint Raymond");
+        HashMap<Short,String> listeSites = new HashMap<Short, String>();
+        listeSites. put((short)1, "Georges Labit");
+        listeSites.put((short)2,"Museum histoire naturelle");
+        listeSites.put((short)3, "Saint Raymond");
         
         try {
            org.omg.CORBA.ORB orb = org.omg.CORBA.ORB.init(args,null);
